break commands out
